+++ libraries/chain/include/graphene/chain/asset_object.hpp
@@ -82,11 +82,11 @@ namespace graphene { namespace chain {
          asset_id_type        asset_id;
          Interest             nominal_interest_perday; //the nominal interest set by asset issuer
          Interest             active_interest_rate;  //active interest = nominal interest ,if interest pool is enough
-         uint16_t             reward_coefficient;
+         uint32_t             reward_coefficient;
          share_type           interest_pool;
          uint32_t             max_period = 2 * JRC_INTEREST_DAYS_YEAR; // as days
-         uint16_t             profile_scale_percent = GRAPHENE_100_PERCENT;// to fit the profile as user wanted at early beginning when interest pool is not enough.
-         uint16_t             profile_receive_percent = GRAPHENE_100_PERCENT;// if the interest pool can't pay for locking balance,user will get the percent of pool balance.
+         uint32_t             profile_scale_percent = GRAPHENE_100_PERCENT;// to fit the profile as user wanted at early beginning when interest pool is not enough.
+         uint32_t             profile_receive_percent = GRAPHENE_100_PERCENT;// if the interest pool can't pay for locking balance,user will get the percent of pool balance.
          bool				  can_unlock_not_expired = true;
       
          share_type    get_profit(share_type tolocking_balance,uint32_t lock_period,const database &_db)const;
